requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pdfplumber==0.10.3
langchain==0.1.0
langchain-openai==0.0.2
langchain-chroma==0.1.0
openai==1.6.1
chromadb==0.4.22
tiktoken==0.5.2
gunicorn==21.2.0